requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
pyjwt==2.10.1
httpx==0.28.1
langchain==0.3.21
SQLAlchemy==2.0.29
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.4
pytest==8.3.5
ai-server-sdk==0.0.20
ecs-logging==2.2.0
pydantic-settings==2.8.1
couchbase==4.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.0
asyncmy==0.2.10
qgdiag-lib-arquitectura